+++ source/Nevermore/SelectBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Nevermore.QueryGraph;
+using Nevermore.AST;
 
 namespace Nevermore
 {
