+++ source/Nevermore/UnaryWhereParameter.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using Nevermore.AST;
 
 namespace Nevermore
 {
